Add files via upload
If you are taking care of the children at home, then this can be a full-time endeavor. To succeed in an online business, you will need to dedicate time to it. So, the first thing that you need to do is to work out how you can find chunks of time to learn all that you need to know about online business and take action to make it happen.

You really do need the support of your family and friends here. If you have a partner, then sit down with them and tell them that you really want to start a successful online business and that you will need to work on it every day. This means that they will need to help you around the home and in the care of the children.